server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
nltk==3.9.1
wikipedia==1.4.0
pydantic==2.9.1
